Added 4 functions: toggle LEDs and print letters
togglered: red led comes on or goes off
toggleblue: blue led comes on or goes off
togglegreen: green led comes on or goes off

printer p: the letter p is printed in hashtags

ex: ##
     #  #
     ##
     #
     #

Note: the typing is acting weird and is not responsive. It will take like 5 presses of each key to make one letter show up in putty. This could be an issue with my terminal though.